👷 parallelize e2e job across browsers via matrix (#4572)
