+++ package.json
@@ -30,13 +30,8 @@
     "test:script": "node --test --experimental-test-module-mocks './scripts/**/*.spec.*'",
     "test:unit:watch": "yarn test:unit --no-single-run",
     "test:unit:bs": "node --env-file-if-exists=.env ./scripts/test/bs-wrapper.ts karma start test/unit/karma.bs.conf.js",
-    "test:e2e:init": "yarn build && yarn build:apps && yarn playwright install chromium --with-deps",
-    "test:e2e": "playwright test --config test/e2e/playwright.local.config.ts --project chromium",
-    "test:e2e:pinned:init": "yarn dlx -p playwright@1.40.1 playwright install firefox webkit",
-    "test:e2e:pinned": "playwright test --config test/e2e/playwright.pinned.config.ts",
-    "test:e2e:bs": "node --env-file-if-exists=.env ./scripts/test/bs-wrapper.ts playwright test --config test/e2e/playwright.bs.config.ts",
-    "test:e2e:ci": "yarn build && yarn build:apps && yarn test:e2e",
-    "test:e2e:ci:bs": "yarn build && yarn build:apps && yarn test:e2e:bs",
+    "test:e2e:init": "yarn build && yarn build:apps && yarn playwright install --with-deps && yarn dlx -p playwright@1.40.1 playwright install firefox webkit",
+    "test:e2e": "playwright test --config test/e2e/playwright.config.ts",
     "test:compat:tsc": "node scripts/check-typescript-compatibility.ts",
     "test:compat:ssr": "scripts/cli check_server_side_rendering_compatibility",
     "test:performance": "yarn test:e2e:init && playwright test --config test/performance/playwright.config.ts",
